Fix check for selection and rename to context-menu rather than right-click-menu
