+++ assets/src/stories-editor/components/context-menu/outside-click-checker.js
