+++ assets/src/stories-editor/components/index.js
@@ -16,7 +16,7 @@ export { default as PageInserter } from './page-inserter';
 export { default as Autocomplete } from './autocomplete';
 export { default as FontFamilyPicker } from './font-family-picker';
 export { default as RotatableBox } from './rotatable-box';
-export { default as RightClickMenu } from './right-click-menu';
+export { default as ContextMenu } from './context-menu';
 export { default as PreviewPicker } from './preview-picker';
 export { default as Inserter } from './inserter';
 export { default as MediaInserter } from './media-inserter';
@@ -29,7 +29,7 @@ export { default as withBlockName } from './higher-order/with-block-name';
 export { default as withCroppedFeaturedImage } from './with-cropped-featured-image';
 export { default as withHasSelectedInnerBlock } from './higher-order/with-has-selected-inner-block';
 export { default as withPageNumber } from './higher-order/with-page-number';
-export { default as withRightClickHandler } from './higher-order/with-right-click-handler';
+export { default as withContextMenu } from './higher-order/with-context-menu';
 export { default as withStoryFeaturedImageNotice } from './higher-order/with-story-featured-image-notice';
 export { default as withEditFeaturedImage } from './with-edit-featured-image';
 export { default as withCustomVideoBlockEdit } from './with-custom-video-block-edit';
