+++ assets/src/stories-editor/index.js
@@ -38,7 +38,7 @@ import {
 	withStoryBlockDropZone,
 	withCallToActionValidation,
 	withCroppedFeaturedImage,
-	withRightClickHandler,
+	withContextMenu,
 } from './components';
 import {
 	maybeEnqueueFontStyle,
@@ -316,7 +316,7 @@ addFilter( 'blocks.registerBlockType', 'ampStoryEditorBlocks/filterBlockTransfor
 addFilter( 'blocks.registerBlockType', 'ampStoryEditorBlocks/deprecateCoreBlocks', deprecateCoreBlocks );
 addFilter( 'editor.BlockEdit', 'ampStoryEditorBlocks/addStorySettings', withAmpStorySettings );
 addFilter( 'editor.BlockEdit', 'ampStoryEditorBlocks/addPageNumber', withPageNumber );
-addFilter( 'editor.BlockEdit', 'ampStoryEditorBlocks/rightClickHandler', withRightClickHandler );
+addFilter( 'editor.BlockEdit', 'ampStoryEditorBlocks/contextMenuHandler', withContextMenu );
 addFilter( 'editor.BlockEdit', 'ampStoryEditorBlocks/addEditFeaturedImage', withEditFeaturedImage );
 addFilter( 'editor.BlockEdit', 'ampEditorBlocks/addVideoBlockPreview', withCustomVideoBlockEdit, 9 );
 addFilter( 'editor.PostFeaturedImage', 'ampStoryEditorBlocks/addFeaturedImageNotice', withStoryFeaturedImageNotice );
